azure_advisor_reports/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
msal==1.24.1

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Processing
pandas==2.1.1
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.6
Pillow==10.1.0

# Cloud Storage
azure-storage-blob==12.19.0

# Environment & Config
python-decouple==3.8
python-dotenv==1.0.0

# Monitoring & Logging
sentry-sdk==1.38.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.10.1
isort==5.12.0
flake8==6.1.0